requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==21.2.0
whitenoise==6.6.0

